Check Desc
I have decided that the best approach would be to develop the front-end first as it is the area I am most familiar with. To do this I will use a created web-design environment in VSCode.

Once the front-end is done to a certain level then I can begin to meddle with the back-end C#. I am not rushing myself on this and am taking notes along the way.