Next steps
Comments detail the next steps in part but for the next while I'll probably only be making visual changes and slight technical changes to the data input.

Advice to myself - stop adding to the project. Stick to what you have now and get past the finish line to get to the next project.